Add Register Structural Descriptions Refactor functions to use new structs. Update enum indexing.

+++ Test/SM72445_AnalogueChannels.test.cpp
@@ -12,10 +12,10 @@ TEST_F(SM72445_Test, getAnalogueChannelAdcResultsNormallyReturnsValue) {
 	EXPECT_CALL(i2c, read(_, Eq(MemoryAddress::REG0))).WillOnce(Return(0x0123'4567'89AB'CDEFull));
 
 	auto adcResults = sm72445.getAnalogueChannelAdcResults().value();
-	EXPECT_EQ(adcResults[static_cast<uint8_t>(AnalogueChannel::CH0)], 0x01EFu);
-	EXPECT_EQ(adcResults[static_cast<uint8_t>(AnalogueChannel::CH2)], 0x02F3u);
-	EXPECT_EQ(adcResults[static_cast<uint8_t>(AnalogueChannel::CH4)], 0x009Au);
-	EXPECT_EQ(adcResults[static_cast<uint8_t>(AnalogueChannel::CH6)], 0x019Eu);
+	EXPECT_EQ(adcResults[AnalogueChannel::CH0], 0x01EFu);
+	EXPECT_EQ(adcResults[AnalogueChannel::CH2], 0x02F3u);
+	EXPECT_EQ(adcResults[AnalogueChannel::CH4], 0x009Au);
+	EXPECT_EQ(adcResults[AnalogueChannel::CH6], 0x019Eu);
 }
 
 TEST_F(SM72445_Test, getAnalogueChannelAdcResultsReturnsNulloptIfI2CReadFails) {
